Update course-shell.md - correct misuse of shebang
In lines 368-376 you mention the first line if the file and talk about how Bash treats the shebang. 
Unfortunately /bin/sh is the POSIX compliant Bourne shell, Bash is the Bourne Again SHell and has extensions to the Bourne shell that do not exist in the Bourne shell. This corrects that.