Update data-wrangling.md -- clarify wording
I initially re-read the "sum of the difference between the two columns" a few times. Subtracting each row individually across the columns and then summing is equivalent to summing each column and then subtracting, so I figured changing the wording could be useful for others. Let me know if I am misunderstanding the exercise provided.

the sum of the difference between the two columns
-->
the difference of the sum of each column